project/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
mediapipe
easyocr
transformers
torch